% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_critical_speed}
\alias{calc_critical_speed}
\title{Calculate Critical Speed and D'}
\usage{
calc_critical_speed(pr_data)
}
\arguments{
\item{pr_data}{a dataframe containing PR performances. The df needs 3 columns.
Columns must be named "athlete", "distance_meters", and "seconds".
Need to convert time to seconds}
}
\value{
a nested dataframe
}
\description{
Calculate Critical Speed and D'
}
