% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_d_balance}
\alias{calc_d_balance}
\title{Calculate D' Balance for runner in race}
\usage{
calc_d_balance(
  runner_data,
  race_data,
  runner_col_names = FALSE,
  race_col_names = FALSE
)
}
\arguments{
\item{runner_data}{a datafra,e containing runners critical speed and D' prime values.}

\item{race_data}{a dataframe containing details of the race to simulate where each row represents a lap split. Needs lap speed & lap distance columns.}
}
\value{

}
\description{
Calculate D' Balance for runner in race
}
