% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_d_balance}
\alias{plot_d_balance}
\title{Plot D' Balance}
\usage{
plot_d_balance(d_balance_data)
}
\arguments{
\item{d_balance_data}{a df. Output from calc_d_balance}
}
\value{
a ggplot
}
\description{
Plot D' Balance
}
