% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_lollipop_race}
\alias{plot_lollipop_race}
\title{Plot Lollipop Race}
\usage{
plot_lollipop_race(runner_data, race_data)
}
\arguments{
\item{runner_data}{a df of runners name, cs, and d' prime.}

\item{race_data}{a df of race data including lap speed and lap distances}
}
\value{
a ggplot
}
\description{
Plot Lollipop Race
}
