% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{to_seconds}
\alias{to_seconds}
\title{Convert MM:SS string to numerical seconds}
\usage{
to_seconds(t)
}
\arguments{
\item{t}{a character string of MM:SS time}
}
\value{
a number in seconds.
}
\description{
Convert MM:SS string to numerical seconds
}
